WEB-3359: PR review fixes
